+++ src/components/cobe/RotateDraggable.tsx
@@ -4,13 +4,12 @@ import { useSpring } from "@react-spring/web";
 import { motion } from "framer-motion";
 import { useCobeStore } from "@/context/Cobe";
 
-export function DraggableCobe() {
+export function RotateDraggableCobe() {
   const { markers } = useCobeStore();
 
   const canvasRef: any = useRef();
   const pointerInteracting: any = useRef(null);
-  const pointerInteractionMovement = useRef(0);
-
+  const pointerInteractionMovement: any = useRef(0);
   const [{ r }, api] = useSpring(() => ({
     r: 0,
     config: {
@@ -20,8 +19,8 @@ export function DraggableCobe() {
       precision: 0.001,
     },
   }));
-
   useEffect(() => {
+    let phi = 0;
     let width = 0;
     const onResize = () =>
       canvasRef.current && (width = canvasRef.current.offsetWidth);
@@ -42,14 +41,20 @@ export function DraggableCobe() {
       glowColor: [1.2, 1.2, 1.2],
       markers,
       onRender: (state) => {
-        state.phi = r.get();
+        // This prevents rotation while dragging
+        if (!pointerInteracting.current) {
+          // Called on every animation frame.
+          // `state` will be an empty object, return updated params.
+          phi += 0.005;
+        }
+        state.phi = phi + r.get();
         state.width = width * 2;
         state.height = width * 2;
       },
     });
     setTimeout(() => (canvasRef.current.style.opacity = "1"));
     return () => globe.destroy();
-  }, [r]);
+  }, [r, markers]);
 
   return (
     <div
